Organize class
